Enviando Listas cpp02 Finalizado

+++ cpp02/ex02/src/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: rbutzke <rbutzke@student.42so.org.br>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/23 09:09:20 by rbutzke           #+#    #+#             */
-/*   Updated: 2024/07/24 10:32:29 by rbutzke          ###   ########.fr       */
+/*   Updated: 2024/07/29 13:32:03 by rbutzke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,49 +16,28 @@
 
 const int	Fixed::_bits = 8;
 
-
 //////////////////////Constructor///////////////////////////
-Fixed::Fixed(): _value(0)
-{	
-	//std::cout << "Default constructor called" << std::endl;
-}
+Fixed::Fixed(): _value(0) {}
 
-Fixed::Fixed(const int add_int): _value(add_int << _bits)
-{	
-	//std::cout << "Int constructor called" << std::endl;
-}
+Fixed::Fixed(const int add_int): _value(add_int << _bits) {}
 
-Fixed::Fixed(const float add_float): _value(roundf(add_float * (1 << _bits)))
-{	
-	//std::cout << "Float constructor called" << std::endl;
-}
+Fixed::Fixed(const float add_float): _value(roundf(add_float * (1 << _bits))){}
 
-Fixed::Fixed(const Fixed &origin)
-{
-	//std::cout << "Copy constructor called" << std::endl;
-	*this = origin;
-}
-
+Fixed::Fixed(const Fixed &origin) {*this = origin;}
 
 //////////////////////Overload copy assignment operator///////////
 Fixed &Fixed::operator=(const Fixed &origin)
 {
-	//std::cout << "Copy assignment operator called" << std::endl;
 	this->setRawBits(origin._value);
 	return (*this);
 }
 
 //////////////////////Destructor methods///////////////////////////
-Fixed::~Fixed()
-{	
-	//std::cout << "Destructor called" << std::endl;
-}
-
-
+Fixed::~Fixed() {}
 
 //////////////////////Conversion methods///////////////////////////
 float Fixed::toFloat() const
-{	return ((float)_value / (float)(1 << Fixed::_bits));}
+{	return ((float)_value / (1 << Fixed::_bits));}
 
 int Fixed::toInt() const
 {	return (_value >> _bits); }
@@ -70,15 +49,12 @@ int Fixed::getRawBits() const
 void Fixed::setRawBits(int const raw)
 {	_value = raw; }
 
-
-
 ////////////////////// Overload insertion operator ///////////////////////////
 std::ostream &operator<<(std::ostream &stream, const Fixed &ref)
 {
 	stream << ref.toFloat();
 	return (stream);
 }
-
 
 ////////////////////// Overload Comparison operator ///////////////////////////
 bool Fixed::operator<(Fixed & ref) const 
@@ -99,10 +75,7 @@ bool Fixed::operator==(Fixed & ref) const
 bool Fixed::operator!=(Fixed & ref) const
 {	return (this->_value != ref.getRawBits());}
 
-
-
 ////////////////////// Overload Arithmetic operator ///////////////////////////
-
 Fixed Fixed::operator-(const Fixed & ref) const
 {	return (Fixed(this->toFloat() - ref.toFloat()));}
 
@@ -114,7 +87,6 @@ Fixed Fixed::operator*(const Fixed & ref) const
 
 Fixed Fixed::operator/(const Fixed & ref) const
 {	return (Fixed(this->toFloat() / ref.toFloat()));}
-
 
 ////////////////////// Incremente decrement prefix ///////////////////////////
 Fixed& Fixed::operator--()
@@ -129,7 +101,7 @@ Fixed& Fixed::operator++()
 	return (*this);
 }
 
-////////////////////// Incremente decrement postfix ///////////////////////////
+////////////////////// Incremente decrement postfix //////////////////////
 Fixed Fixed::operator--(int)
 {
 	Fixed temp(*this);
@@ -145,7 +117,6 @@ Fixed Fixed::operator++(int)
 	this->_value++;
 	return (temp);
 }
-
 
 ////////////////////////static member function///////////////////////////
 Fixed &Fixed::min(Fixed & ref_A, Fixed & ref_B)
